Now we just resume from sleep via WDT
Conditions that cause the contactor to set:
a) boot up
b) contactor requested to change states
c) periodicly based on wakes from sleep

Contactor state is not stored to EEPROM. Not needed because this PIC
doesn't reboot on WDT from sleep. It just resumes. So all variables
are maintained. I was previously using EEPROM because most PICs (used
to) essentially cold boot on WDT timeout.